[FIX] PHP 7.3: Unparenthesized expressions
PHP 7.3 raises a deprecation notice when the concat operator (.) and +/- operators are used in the same expression without parenthesis. This is to mitigate the code ambiguity, and to prepare to set the strict order of operators. From PHP 8 and later, this will be evaluated as 'total' + (5 + 5), with the + and - operators taking higher precedence.